- add share content for both death and beasts
